Shuffle some code around.
