--- agent/balrogagent/telemetry.py
@@ -1,2 +1,0 @@
-def get_telemetry_uptake(uri, product, channel):
-    pass
